Start of development: Version 2.0.5 beta 1
- FDISK now has its own printf

- ROM files generated renamed to v2.0.5 beta 1

- Added gitignores in all the bank folders

+++ source/msxdos25/bank5/AsmCall.c
@@ -30,7 +30,7 @@ void DriverCall(byte slot, uint routineAddress)
 void DosCall(byte function, register_usage outRegistersDetail)
 {
     regs.Bytes.C = function;
-    SwitchSystemBankThenCall(0xF37D, outRegistersDetail);
+    SwitchSystemBankThenCall((int)0xF37D, outRegistersDetail);
 }
